Lunarg push_constants sample, made void-return vk methods return void from driver

+++ command_barrier.go
@@ -42,8 +42,8 @@ type BufferMemoryBarrierOptions struct {
 
 	Buffer Buffer
 
-	Offset uint64
-	Size   uint64
+	Offset int
+	Size   int
 
 	common.HaveNext
 }
